nethelper, logger moved to outside of hcollections module, references to it changed, exceptions rewrapped
