+++ python2.7libs/hpaste/logger.py
